--- bluetooth/ble.py
@@ -1,2 +1,0 @@
-from gattlib import *
-
